feat: move sidebar to separate dir
